moves locker interface and instance to it's own file
